fixed: emacs getting stuck
org-habit was a org-module not a package,
to enable it, we needed to configure it.

this is now set in 'custom.el', which ideally
should be migrated into 'config.el' but it's working
now.

also as a byproduct, most of the settings is in after!
meaning should be better in terms of contaminating
global namespace, or potentially speed up (since it's lazy??)